Move @internal classes to Io namespace
