+++ src/Io/ChunkedDecoder.php
@@ -1,10 +1,11 @@
 <?php
-namespace React\Http;
+
+namespace React\Http\Io;
 
 use Evenement\EventEmitter;
 use React\Stream\ReadableStreamInterface;
-use React\Stream\WritableStreamInterface;
 use React\Stream\Util;
+use React\Stream\WritableStreamInterface;
 use Exception;
 
 /** @internal */
