+++ src/Io/ChunkedEncoder.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace React\Http;
+namespace React\Http\Io;
 
 use Evenement\EventEmitter;
 use React\Stream\ReadableStreamInterface;
-use React\Stream\WritableStreamInterface;
 use React\Stream\Util;
+use React\Stream\WritableStreamInterface;
 
 /** @internal */
 class ChunkedEncoder extends EventEmitter implements ReadableStreamInterface
