+++ src/Io/CloseProtectionStream.php
@@ -1,13 +1,14 @@
 <?php
 
-namespace React\Http;
+namespace React\Http\Io;
 
 use Evenement\EventEmitter;
 use React\Stream\ReadableStreamInterface;
-use React\Stream\WritableStreamInterface;
 use React\Stream\Util;
+use React\Stream\WritableStreamInterface;
 
-/** @internal
+/**
+ * @internal
  * This stream is used to protect the passed stream against closing.
  * */
 class CloseProtectionStream extends EventEmitter implements ReadableStreamInterface
