+++ src/Io/LengthLimitedStream.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace React\Http;
+namespace React\Http\Io;
 
 use Evenement\EventEmitter;
 use React\Stream\ReadableStreamInterface;
-use React\Stream\WritableStreamInterface;
 use React\Stream\Util;
+use React\Stream\WritableStreamInterface;
 
 /** @internal */
 class LengthLimitedStream extends EventEmitter implements ReadableStreamInterface
