+++ src/Server.php
@@ -6,6 +6,13 @@ use Evenement\EventEmitter;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use React\Http\Io\ChunkedDecoder;
+use React\Http\Io\ChunkedEncoder;
+use React\Http\Io\CloseProtectionStream;
+use React\Http\Io\HttpBodyStream;
+use React\Http\Io\LengthLimitedStream;
+use React\Http\Io\RequestHeaderParser;
+use React\Http\Io\ServerRequest;
 use React\Promise\CancellablePromiseInterface;
 use React\Promise\Promise;
 use React\Socket\ConnectionInterface;
